Settings page almost complete
The picture is missing but it's not essential. The list of profiles and the input fields are generated from the content Front Matter. The fields can be edited and information will be retained until the page is reloaded.

The settings page is bilingual FR/EN, but selecting another language does not yet switch it.