Merge 7e0d8c90 (#5)
* Initial commit

* first commit

* Settings page almost complete

The picture is missing but it's not essential. The list of profiles and the input fields are generated from the content Front Matter. The fields can be edited and information will be retained until the page is reloaded.

The settings page is bilingual FR/EN, but selecting another language does not yet switch it.

* 🔨 Add git_toolbox_prj.xml to .gitignore

* ✨🔧 Add og:image to single.html

- Added og:image meta tag in single.html
- Updated the list of defects in README.md

* Hello World by Anima

* Add build script for Render.com

* Made directory changes temporary in build.sh

* Use npm instead of yarn to build react app

* Update parcel to latest

* Add yarn install to install node dependencies

* Update README.md

Translated into English

* Update README.md

---------

Co-authored-by: Jean-Marc Le Peuvedic <26287656+lepeuvedic@users.noreply.github.com>
Co-authored-by: Jean-Marc Le Peuvédic <lepeuvedic@gmail.com>